requirements.txt
Django==4.2.13
djangorestframework==3.15.2
psycopg2-binary==2.9.10
celery==5.3.6
redis==5.0.3
django-cors-headers==4.3.1
gunicorn==21.2.0
python-dotenv==1.0.1
